Refactor task settings rows to share a single state source
